services/producer/requirements.txt
kafka-python==2.0.2
APScheduler==3.10.4
prometheus-client==0.19.0
numpy==1.26.4
pandas==2.2.2
yfinance>=1.0.0
